+++ activemodel/CHANGELOG.md
@@ -1,2 +1,6 @@
+*   Allow string types to observe the `limit` parameter and chop the
+    resulting string to the specified number of characters.
+
+    *Stéphane Maniaci*
 
 Please check [7-2-stable](https://github.com/rails/rails/blob/7-2-stable/activemodel/CHANGELOG.md) for previous changes.